Add ability to correct the score of teams at any point.
This commit adds a correct score button that will display a form with a
team # and score change box. The specified team's score will be adjusted
by the amount in the score change box. To remove 100 points from a team,
enter -100 in the score change box. To add 200 points to a team, enter
200 in the score change box.